Fix marker & waveform alignment (OT-848) (#1071)

+++ jvm/markerapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/markerapp/app/viewmodel/VerseMarkerViewModel.kt
@@ -287,8 +287,8 @@ class VerseMarkerViewModel : ViewModel(), IMarkerViewModel {
         }
 
         val samplesPerScreenWidth = sampleRate * secondsOnScreen
-        val samplesPerPixel = samplesPerScreenWidth / width.toDouble()
+        val samplesPerPixel = samplesPerScreenWidth / width
         val pixelsInDuration = waveformAudioPlayerProperty.get().getDurationInFrames() / samplesPerPixel
-        return pixelsInDuration
+        return pixelsInDuration.toDouble()
     }
 }
